requirements.txt
Flask==1.0.2
requests==2.21.0
cython==0.29.7
cassandra-driver==3.17.0
numpy==1.16.3
pandas==0.24.2
python-dateutil==2.8.0
ndjson==0.1.0
